requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
dashscope==1.22.2
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
propcache==0.3.0
python-dotenv==1.0.1
python-telegram-bot==21.11.1
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
yarl==1.18.3
